Add files via upload
In the new folder, HW5 working, I added the changed HW5sol.hs file as well as copies of the verifier and types files. In HW5sol line 55 I changed the run function so that it first sent a RankError if isNothing is true. otherwise it calls the helper function. I essentially switched the calls at line 55 and 56.